wip: clear analyzer quality findings
